requirements.txt
accelerate
deepspeed==0.18.5
mujoco==3.4.0
numpy==1.26.4
pytorch-lightning==2.6.1
tokenizers==0.13.3
torchmetrics==1.8.2
transformers==4.31.0

# 最新版本（自动更新）
absl-py
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
async-timeout
attrs
certifi
charset-normalizer
click
contourpy
cycler
decorator
decord
easydict
egl_probe
einops
etils
evdev
exceptiongroup
filelock
flow-vis
fonttools
frozenlist
fsspec
gitdb
GitPython
glfw
grpcio
h11
h5py
hf-xet
httpcore
httpx
huggingface-hub
hydra-core
idna
ImageIO
imageio-ffmpeg
importlib_resources
iniconfig
Jinja2
kiwisolver
lightning
lightning-utilities
llvmlite
Markdown
MarkupSafe
matplotlib
modelscope
mink
moviepy==1.0.3
mpmath
msgpack
multidict
natsort
networkx
numba
nvitop
omegaconf
opencv-python
packaging
pandas
pillow
pip
platformdirs
pluggy
proglog
propcache
protobuf
psutil
pyarrow
pydantic
pydantic_core
Pygments
pynput
PyOpenGL
pyparsing
pytest
python-dateutil
python-dotenv
python-xlib
pytz
PyYAML
qpsolvers
quadprog
regex
requests
safetensors
scipy
sentry-sdk
setuptools
shellingham
six
smmap
sympy
tensorboard
tensorboard-data-server
tensorboardX
termcolor
timm
tomli
tqdm
typer-slim
typing-inspection
typing_extensions
tzdata
urllib3
wandb
Werkzeug
wheel
yarl
zipp
